Make the PDF outline links active in Adobe Reader
The /FitH entry of the /View key requires a "top" parameter, so this usage was
invalid. It worked with Okular, which uses the Poppler PDF library.

The destination of the three "Manual pages" headings now is near the end
of the previous page, which does not look good, but I couldn't do
better.